Moves parseInboundCommand into Slack lib
